Move out inner class and remove interface
